+++ bukkit/src/main/java/com/github/games647/changeskin/bukkit/tasks/NameResolver.java
@@ -3,11 +3,9 @@ package com.github.games647.changeskin.bukkit.tasks;
 import com.github.games647.changeskin.bukkit.ChangeSkinBukkit;
 import com.github.games647.changeskin.core.NotPremiumException;
 import com.github.games647.changeskin.core.RateLimitException;
-import com.github.games647.changeskin.core.model.SkinData;
 
 import java.util.UUID;
 import java.util.logging.Level;
-import org.bukkit.Bukkit;
 
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -38,12 +36,6 @@ public class NameResolver implements Runnable {
                 return;
             }
 
-            SkinData targetSkin = plugin.getStorage().getSkin(targetName);
-            if (targetSkin != null) {
-                onNameResolveDatabase(targetSkin);
-                return;
-            }
-
             try {
                 uuid = plugin.getCore().getMojangSkinApi().getUUID(targetName);
                 if (uuid == null) {
@@ -52,7 +44,6 @@ public class NameResolver implements Runnable {
                     }
                 } else {
                     plugin.getCore().getUuidCache().put(targetName, uuid);
-                    onNameResolve(uuid);
                 }
             } catch (NotPremiumException notPremiumEx) {
                 plugin.getLogger().log(Level.FINE, "Requested not premium", notPremiumEx);
@@ -67,23 +58,11 @@ public class NameResolver implements Runnable {
                     plugin.sendMessage(invoker, "rate-limit");
                 }
             }
-        } else {
+        }
+
+        if (uuid != null) {
             onNameResolve(uuid);
         }
-    }
-
-    private void onNameResolveDatabase(SkinData targetSkin) {
-        UUID uuid = targetSkin.getUuid();
-        if (invoker != null) {
-            plugin.sendMessage(invoker, "uuid-resolved");
-            if (plugin.getConfig().getBoolean("skinPermission") && !plugin.checkPermission(invoker, uuid, true)) {
-                return;
-            }
-
-            plugin.sendMessage(invoker, "skin-downloading");
-        }
-
-        Bukkit.getScheduler().runTask(plugin, new SkinUpdater(plugin, invoker, player, targetSkin));
     }
 
     private void onNameResolve(UUID uuid) {