+++ bungee/src/main/java/com/github/games647/changeskin/bungee/listener/AbstractSkinListener.java
@@ -27,13 +27,6 @@ public abstract class AbstractSkinListener implements Listener {
     public void refetch(UserPreference preferences, String playerName) {
         UUID ownerUUID = plugin.getCore().getUuidCache().get(playerName);
         if (ownerUUID == null && !plugin.getCore().getCrackedNames().containsKey(playerName)) {
-            SkinData skin = plugin.getStorage().getSkin(playerName);
-            if (skin != null) {
-                preferences.setTargetSkin(skin);
-                save(skin, preferences);
-                return;
-            }
-
             try {
                 ownerUUID = plugin.getCore().getMojangSkinApi().getUUID(playerName);
             } catch (NotPremiumException ex) {