+++ sponge/src/main/java/com/github/games647/changeskin/sponge/tasks/NameResolver.java
@@ -2,7 +2,6 @@ package com.github.games647.changeskin.sponge.tasks;
 
 import com.github.games647.changeskin.core.NotPremiumException;
 import com.github.games647.changeskin.core.RateLimitException;
-import com.github.games647.changeskin.core.model.SkinData;
 import com.github.games647.changeskin.sponge.ChangeSkinSponge;
 
 import java.util.UUID;
@@ -36,12 +35,6 @@ public class NameResolver implements Runnable {
                 return;
             }
 
-            SkinData targetSkin = plugin.getCore().getStorage().getSkin(targetName);
-            if (targetSkin != null) {
-                onNameResolveDatabase(targetSkin);
-                return;
-            }
-
             try {
                 uuid = plugin.getCore().getMojangSkinApi().getUUID(targetName);
                 if (uuid == null) {
@@ -50,7 +43,6 @@ public class NameResolver implements Runnable {
                     }
                 } else {
                     plugin.getCore().getUuidCache().put(targetName, uuid);
-                    onNameResolve(uuid);
                 }
             } catch (NotPremiumException notPremiumEx) {
                 plugin.getLogger().debug("Requested not premium", notPremiumEx);
@@ -65,24 +57,11 @@ public class NameResolver implements Runnable {
                     plugin.sendMessage(invoker, "rate-limit");
                 }
             }
-        } else {
+        }
+
+        if (uuid != null) {
             onNameResolve(uuid);
         }
-    }
-
-    private void onNameResolveDatabase(SkinData targetSkin) {
-        if (invoker != null) {
-            plugin.sendMessage(invoker, "uuid-resolved");
-            if (plugin.getRootNode().getNode("skinPermission").getBoolean()
-                    && !plugin.checkPermission(invoker, targetSkin.getUuid(), true)) {
-                return;
-            }
-
-            plugin.sendMessage(invoker, "skin-downloading");
-        }
-
-        SkinUpdater skinUpdater = new SkinUpdater(plugin, invoker, receiver, targetSkin);
-        plugin.getGame().getScheduler().createTaskBuilder().execute(skinUpdater).submit(plugin);
     }
 
     private void onNameResolve(UUID uuid) {